langIt.txt (#24)
Proposed the first step of the Italian translation. In some cases I have kept the sentences as short as possible, so there are different adaptations from the literal translation. Also, I had some problems with accented words (such as words with characters), so I opted for apostrophes. I hope this file can be suitable.